12+ HOURS OF VARIOUS CODING. Hopefully backend is done.

+++ src/main/java/edu/redwoods/cis18/assemble/repository/EventRepository.java
@@ -3,7 +3,5 @@ package edu.redwoods.cis18.assemble.repository;
 import edu.redwoods.cis18.assemble.model.Event;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-// EventRepository extends JpaRepository, providing built-in methods for CRUD operations
-public interface EventRepository extends JpaRepository<Event, Integer> {
-    // Additional custom queries can be added here if needed
+public interface EventRepository extends JpaRepository<Event, Long> {
 }
